% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the daedalus.shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\value{
A \code{shiny.appobj} (invisibly). Called for side-effects.
}
\description{
Launches the Shiny dashboard for configuring and running DAEDALUS
epidemiological-economic model scenarios.
}
